The Standard SQL names "win" because Standard SQL is where the future
is. We need to have Array in the type for parameters, but it's not how
repeated fields are represented in schemas... this is probably a
reasonable compromise for now.

+++ apis/Google.Cloud.BigQuery.V2/Google.Cloud.BigQuery.V2.Snippets/BigQuerySnippetFixture.cs
@@ -56,8 +56,8 @@ namespace Google.Cloud.BigQuery.V2.Snippets
             var historySchema = new TableSchemaBuilder
             {
                 { "player", BigQueryDbType.String },
-                { "score", BigQueryDbType.Integer },
-                { "level", BigQueryDbType.Integer },
+                { "score", BigQueryDbType.Int64 },
+                { "level", BigQueryDbType.Int64 },
                 { "game_started", BigQueryDbType.Timestamp }
             }.Build();
             var historyTable = game.CreateTable(HistoryTableId, historySchema);
